Changed long to BigDecimal

+++ src/model/Purchase.java
@@ -1,20 +1,22 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Purchase {
     private String myName;
-    private double myPrice;
+    private BigDecimal myPrice;
     private String myDate;
 
-    public Purchase(String theName, double thePrice, String theDate) {
+    public Purchase(String theName, BigDecimal thePrice, String theDate) {
         myName = theName;
         myPrice = thePrice;
         myDate = theDate;
     }
 
     private void setName(String theName) { myName = theName; }
-    private void setPrice(long thePrice) { myPrice = thePrice; }
+    private void setPrice(BigDecimal thePrice) { myPrice = thePrice; }
     private void setDate(String theDate) {myDate = theDate; }
     public String getName() { return myName; }
-    public double getPrice() { return myPrice; }
+    public BigDecimal getPrice() { return myPrice; }
     public String getDate() { return myDate; }
 }
